-История поиска, добавление, удаление -ProgressBar

+++ app/src/main/java/com/example/surimusakotlin/model/AltMeasure.kt
@@ -2,7 +2,7 @@ package com.example.surimusakotlin.model
 
 data class AltMeasure(
     val measure: String,
-    val qty: Int,
-    val seq: Int,
+    val qty: Double,
+    val seq: Double,
     val serving_weight: Double
 )
